requirements.txt
face_recognition
Pillow
flake8
dlib
rich
FastAPI
uvicorn
python-multipart